requirements.txt
requests
sentry_sdk
posthog
py-machineid
py
python-dotenv